Refactor signal_generator into smaller functions
signal_generator was split into two functions: random_frequencies_generator, that generates random frequencies and signal_generator, that actually produces the signal. Maximum number of componets is now fixed and not randomly extracted for simplicity. The tests in test_signal_generator.py are updated to test the typical usage case of the new signal_generator function.